Absolute positioning, G28, 91, 90 and 92 implemented

+++ src/control/stepper.h
@@ -27,7 +27,7 @@ private:
   gpio_t dir;
   gpio_t enable;
 public:
-  uint16_t pos;
+  double pos;
   uint16_t stepsPerMM;
 
   float micro_delay;
